Use six for Python 2 and 3 urlparse
urlparse in python 2 has been moved to urllib.parse in python 3. Six, a
python 2 and 3 compatibility library allows you to use urlparse in the
python 3 like way for both versions.

I've added six into this repo directly so that it doesn't have to be
installed on the AWS AMIs and will be pulled in via this source (so we
don't have to update the AMI yet).

Eventually, six should be removed here once a packaging/env method is
adopted, or the AMI is updated.