Simplify release action
Not querying the PR for the PR body significantly simplifies the release
workflow and the release body text is unnecessary. We're distributing
releases through a python package index, and keeping changes in our
changelog, so just the title is sufficient.

Also, the draft and prerelease, settings are the defualt, so no need to
include them:
https://github.com/actions/create-release/blob/c9ba6969f07ed90fae07e2e66100dd03f9b1a50e/action.yml#L11-L21