Fix cleaning threshold for power data
Because the cleaning threshold was applied before the conversion from
amplitude to power, the cleaning threshold is applied differently to
these two types.

The hard coded value here was the correct value for amplitude data.
Note: 0.0039811 = pow(10.0, -24.0 / 10.0) and intended for amplitude,
so the corresponding threshold in power is pow(10.0, -48.0 / 10.0).

This converts to power before cleaning, and then applies the
corresponding power cleaning threshold afterwards.